build: add dockerfile to run tests on linux (#22)
It speeds up development to have an option to test changes on Linux when
otherwise developing on Windows or macOS. The `--no-sandbox` argument is
required when launching electron to support running as root in the
container (and has little effect otherwise). Setting ELECTRON_VERSION is
required to build the container as it installs that specific version of
electron which will later be used when running the integration tests.

  Example:

    ELECTRON_VERSION='^11.3.0' docker build --tag pkg .
    docker run --interactive --tty pkg

  Shorthand:

    docker build -t pkg . && docker run -it pkg

  To later remove the image:

    docker rmi pkg

The .dockerignore file avoids copying dependencies and prebuilt files
into the container, which can take significant time particularly for
node_modules.